[BN-1019]: Changed directory strucure, drivers inside driver dir
